requirements.txt
Django==2.1.4
djangorestframework==3.9.0
Markdown==3.0.1
django-filter==2.0.0
django-phonenumber-field==2.1.0
phonenumbers==8.10.2
django-cpffield==0.1.1
Pillow==5.3.0
django-simple-history==2.6.0
pycodestyle==2.4.0
flake8-django==0.0.3
django-rest-swagger==2.2.0
psycopg2==2.7.6.1
psycopg2-binary
